Fix compilation for 32bits
